Revert "Merge pull request #152 from VNG-Realisatie/add-verzending-resource"
This reverts commit 72a0a8f70df9a7c375fd8e41a1b904a050b076c3, reversing
changes made to 870b7b4d6888f3dafe00cf878b661c72c2e21d94.
