+++ bin/compile_dependencies.cmd
@@ -6,12 +6,12 @@ cd %toplevel%
 
 REM Base deps
 pip-compile^
-    --no-emit-index-url^
+    --no-index^
     requirements/base.in
 
 REM Jenkins/tests deps
 pip-compile^
-    --no-emit-index-url^
+    --no-index^
     --output-file requirements/jenkins.txt^
     requirements/base.txt^
     requirements/testing.in^
