+++ requirements/base.txt
@@ -1,185 +1,69 @@
 #
-# This file is autogenerated by pip-compile with python 3.7
+# This file is autogenerated by pip-compile
 # To update, run:
 #
-#    pip-compile --no-emit-index-url requirements/base.in
+#    pip-compile --no-index requirements/base.in
 #
-alabaster==0.7.12
-    # via sphinx
-babel==2.6.0
-    # via sphinx
-certifi==2018.4.16
-    # via requests
-chardet==3.0.4
-    # via requests
-coreapi==2.3.3
-    # via drf-yasg
-coreschema==0.0.4
-    # via
-    #   coreapi
-    #   drf-yasg
-django==2.2.4
-    # via
-    #   -r requirements/base.in
-    #   django-choices
-    #   django-extra-fields
-    #   django-filter
-    #   django-markup
-    #   django-privates
-    #   django-rest-framework-condition
-    #   django-sendfile2
-    #   drf-nested-routers
-    #   drf-yasg
-    #   vng-api-common
-django-appconf==1.0.2
-    # via django-axes
-django-axes==4.4.0
-    # via -r requirements/base.in
-django-choices==1.7.0
-    # via
-    #   -r requirements/base.in
-    #   vng-api-common
-django-cors-middleware==1.3.1
-    # via -r requirements/base.in
-django-extra-fields==2.0.3
-    # via -r requirements/base.in
-django-filter==2.0.0
-    # via
-    #   -r requirements/base.in
-    #   djangorestframework-filters
-    #   vng-api-common
-django-ipware==2.1.0
-    # via django-axes
-django-markup==1.3
-    # via -r requirements/base.in
-django-privates==1.0.1.post0
-    # via -r requirements/base.in
-django-rest-framework-condition==0.1.1
-    # via vng-api-common
-django-sendfile2==0.4.2
-    # via django-privates
-django-solo==1.1.3
-    # via vng-api-common
-djangorestframework==3.9.4
-    # via
-    #   -r requirements/base.in
-    #   django-extra-fields
-    #   djangorestframework-filters
-    #   drf-nested-routers
-    #   drf-yasg
-    #   vng-api-common
-djangorestframework-camel-case==0.2.0
-    # via
-    #   -r requirements/base.in
-    #   vng-api-common
-djangorestframework-filters==0.10.2
-    # via -r requirements/base.in
-docutils==0.14
-    # via sphinx
-drf-nested-routers==0.90.2
-    # via vng-api-common
-drf-writable-nested==0.6.3
-    # via -r requirements/base.in
-drf-yasg==1.16.0
-    # via
-    #   -r requirements/base.in
-    #   vng-api-common
-gemma-zds-client==0.13.4
-    # via vng-api-common
-humanize==0.5.1
-    # via -r requirements/base.in
-idna==2.7
-    # via requests
-imagesize==1.1.0
-    # via sphinx
-inflection==0.3.1
-    # via drf-yasg
-iso-639==0.4.5
-    # via vng-api-common
-isodate==0.6.0
-    # via vng-api-common
-itypes==1.1.0
-    # via coreapi
-jinja2==2.10.1
-    # via
-    #   coreschema
-    #   sphinx
-markdown==3.1
-    # via -r requirements/base.in
-markupsafe==1.1.1
-    # via jinja2
-oyaml==0.7
-    # via vng-api-common
-packaging==19.0
-    # via sphinx
-pillow==6.2.1
-    # via -r requirements/base.in
-psycopg2-binary==2.7.5
-    # via -r requirements/base.in
-pygments==2.3.1
-    # via sphinx
-pyjwt==1.6.4
-    # via
-    #   gemma-zds-client
-    #   vng-api-common
-pyparsing==2.3.1
-    # via packaging
-python-dateutil==2.7.3
-    # via -r requirements/base.in
-python-dotenv==0.8.2
-    # via -r requirements/base.in
-pytz==2018.4
-    # via
-    #   -r requirements/base.in
-    #   babel
-    #   django
-    #   django-axes
-pyyaml==5.1.2
-    # via
-    #   gemma-zds-client
-    #   oyaml
-    #   vng-api-common
-raven==6.9.0
-    # via -r requirements/base.in
-requests==2.21
-    # via
-    #   coreapi
-    #   gemma-zds-client
-    #   sphinx
-    #   vng-api-common
-ruamel.yaml==0.15.40
-    # via drf-yasg
-six==1.11.0
-    # via
-    #   django-markup
-    #   django-sendfile2
-    #   drf-yasg
-    #   isodate
-    #   packaging
-    #   python-dateutil
-    #   sphinx
-snowballstemmer==1.2.1
-    # via sphinx
-sphinx==1.8.4
-    # via
-    #   -r requirements/base.in
-    #   sphinx-rtd-theme
-sphinx-rtd-theme==0.4.2
-    # via -r requirements/base.in
-sphinxcontrib-websupport==1.1.0
-    # via sphinx
-sqlparse==0.3.0
-    # via django
-unidecode==1.0.22
-    # via vng-api-common
-uritemplate==3.0.0
-    # via
-    #   coreapi
-    #   drf-yasg
-urllib3==1.24.3
-    # via requests
-vng-api-common==1.5.10
-    # via -r requirements/base.in
+alabaster==0.7.12         # via sphinx
+babel==2.6.0              # via sphinx
+certifi==2018.4.16        # via requests
+chardet==3.0.4            # via requests
+coreapi==2.3.3            # via drf-yasg
+coreschema==0.0.4         # via coreapi, drf-yasg
+django-appconf==1.0.2     # via django-axes
+django-axes==4.4.0        # via -r requirements/base.in
+django-choices==1.7.0     # via -r requirements/base.in, vng-api-common
+django-cors-middleware==1.3.1  # via -r requirements/base.in
+django-extra-fields==2.0.3  # via -r requirements/base.in
+django-filter==2.0.0      # via -r requirements/base.in, djangorestframework-filters, vng-api-common
+django-ipware==2.1.0      # via django-axes
+django-markup==1.3        # via -r requirements/base.in
+django-privates==1.0.1.post0  # via -r requirements/base.in
+django-rest-framework-condition==0.1.1  # via vng-api-common
+django-sendfile2==0.4.2   # via django-privates
+django-solo==1.1.3        # via vng-api-common
+django==2.2.4             # via -r requirements/base.in, django-choices, django-extra-fields, django-filter, django-markup, django-privates, django-rest-framework-condition, django-sendfile2, drf-nested-routers, drf-yasg, vng-api-common
+djangorestframework-camel-case==0.2.0  # via -r requirements/base.in, vng-api-common
+djangorestframework-filters==0.10.2  # via -r requirements/base.in
+djangorestframework==3.9.4  # via -r requirements/base.in, django-extra-fields, djangorestframework-filters, drf-nested-routers, drf-yasg, vng-api-common
+docutils==0.14            # via sphinx
+drf-nested-routers==0.90.2  # via vng-api-common
+drf-yasg==1.16.0          # via -r requirements/base.in, vng-api-common
+gemma-zds-client==0.13.4  # via vng-api-common
+humanize==0.5.1           # via -r requirements/base.in
+idna==2.7                 # via requests
+imagesize==1.1.0          # via sphinx
+inflection==0.3.1         # via drf-yasg
+iso-639==0.4.5            # via vng-api-common
+isodate==0.6.0            # via vng-api-common
+itypes==1.1.0             # via coreapi
+jinja2==2.10.1            # via coreschema, sphinx
+markdown==3.1             # via -r requirements/base.in
+markupsafe==1.1.1         # via jinja2
+oyaml==0.7                # via vng-api-common
+packaging==19.0           # via sphinx
+pillow==6.2.1             # via -r requirements/base.in
+psycopg2-binary==2.7.5    # via -r requirements/base.in
+pygments==2.3.1           # via sphinx
+pyjwt==1.6.4              # via gemma-zds-client, vng-api-common
+pyparsing==2.3.1          # via packaging
+python-dateutil==2.7.3    # via -r requirements/base.in
+python-dotenv==0.8.2      # via -r requirements/base.in
+pytz==2018.4              # via -r requirements/base.in, babel, django, django-axes
+pyyaml==5.1.2             # via gemma-zds-client, oyaml, vng-api-common
+raven==6.9.0              # via -r requirements/base.in
+requests==2.21            # via coreapi, gemma-zds-client, sphinx, vng-api-common
+ruamel.yaml==0.15.40      # via drf-yasg
+six==1.11.0               # via django-markup, django-sendfile2, drf-yasg, isodate, packaging, python-dateutil, sphinx
+snowballstemmer==1.2.1    # via sphinx
+sphinx-rtd-theme==0.4.2   # via -r requirements/base.in
+sphinx==1.8.4             # via -r requirements/base.in, sphinx-rtd-theme
+sphinxcontrib-websupport==1.1.0  # via sphinx
+sqlparse==0.3.0           # via django
+unidecode==1.0.22         # via vng-api-common
+uritemplate==3.0.0        # via coreapi, drf-yasg
+urllib3==1.24.3           # via requests
+vng-api-common==1.5.10    # via -r requirements/base.in
 
 # The following packages are considered to be unsafe in a requirements file:
 # setuptools
