+++ src/drc/api/apps.py
@@ -6,4 +6,4 @@ class DRCApiConfig(AppConfig):
 
     def ready(self):
         # ensure that the metaclass for every viewset has run
-        from . import views  # noqa
+        from . import viewsets  # noqa
