+++ src/drc/api/filters.py
@@ -8,7 +8,6 @@ from drc.datamodel.models import (
     EnkelvoudigInformatieObjectCanonical,
     Gebruiksrechten,
     ObjectInformatieObject,
-    Verzending,
 )
 
 
@@ -50,20 +49,4 @@ class GebruiksrechtenFilter(FilterSet):
             "informatieobject": ["exact"],
             "startdatum": ["lt", "lte", "gt", "gte"],
             "einddatum": ["lt", "lte", "gt", "gte"],
-        }
-
-
-class VerzendingFilter(FilterSet):
-    informatieobject = URLModelChoiceFilter(
-        queryset=EnkelvoudigInformatieObjectCanonical.objects.all(),
-        instance_path="canonical",
-        help_text=get_help_text("datamodel.Verzending", "informatieobject"),
-    )
-
-    class Meta:
-        model = Verzending
-        fields = {
-            "aard_relatie": ["exact"],
-            "informatieobject": ["exact"],
-            "betrokkene": ["exact"],
         }
