--- src/drc/api/serializers/__init__.py
@@ -1,11 +1,0 @@
-from .bestandsdeel import BestandsDeelSerializer
-from .enkelvoudig_informatieobject import (
-    EnkelvoudigInformatieObjectCreateLockSerializer,
-    EnkelvoudigInformatieObjectSerializer,
-    EnkelvoudigInformatieObjectWithLockSerializer,
-    LockEnkelvoudigInformatieObjectSerializer,
-    UnlockEnkelvoudigInformatieObjectSerializer,
-)
-from .gebruiksrechten import GebruiksrechtenSerializer
-from .object_informatieobject import ObjectInformatieObjectSerializer
-from .verzending import VerzendingSerializer
