+++ src/drc/api/urls.py
@@ -4,13 +4,12 @@ from django.urls import include, path
 from vng_api_common import routers
 from vng_api_common.schema import SchemaView
 
-from .views import (
+from .viewsets import (
     BestandsDeelViewSet,
     EnkelvoudigInformatieObjectAuditTrailViewSet,
     EnkelvoudigInformatieObjectViewSet,
     GebruiksrechtenViewSet,
     ObjectInformatieObjectViewSet,
-    VerzendingViewSet,
 )
 
 router = routers.DefaultRouter()
@@ -23,7 +22,6 @@ router.register(
 router.register("gebruiksrechten", GebruiksrechtenViewSet)
 router.register("objectinformatieobjecten", ObjectInformatieObjectViewSet)
 router.register("bestandsdelen", BestandsDeelViewSet)
-router.register("verzendingen", VerzendingViewSet)
 
 # TODO: the EndpointEnumerator seems to choke on path and re_path
