--- src/drc/api/views/__init__.py
@@ -1,8 +1,0 @@
-from .bestandsdeel import BestandsDeelViewSet
-from .enkelvoudig_informatieobject import (
-    EnkelvoudigInformatieObjectAuditTrailViewSet,
-    EnkelvoudigInformatieObjectViewSet,
-)
-from .gebruiksrechten import GebruiksrechtenViewSet
-from .object_informatieobject import ObjectInformatieObjectViewSet
-from .verzending import VerzendingViewSet
