--- src/drc/api/views/gebruiksrechten.py
@@ -1,87 +1,0 @@
-from rest_framework import viewsets
-from vng_api_common.audittrails.viewsets import AuditTrailViewsetMixin
-from vng_api_common.caching.decorators import conditional_retrieve
-from vng_api_common.notifications.viewsets import NotificationViewSetMixin
-from vng_api_common.viewsets import CheckQueryParamsMixin
-
-from drc.api.audits import AUDIT_DRC
-from drc.api.data_filtering import ListFilterByAuthorizationsMixin
-from drc.api.filters import GebruiksrechtenFilter
-from drc.api.kanalen import KANAAL_DOCUMENTEN
-from drc.api.permissions import InformationObjectRelatedAuthScopesRequired
-from drc.api.scopes import (
-    SCOPE_DOCUMENTEN_AANMAKEN,
-    SCOPE_DOCUMENTEN_ALLES_LEZEN,
-    SCOPE_DOCUMENTEN_ALLES_VERWIJDEREN,
-    SCOPE_DOCUMENTEN_BIJWERKEN,
-)
-from drc.api.serializers import GebruiksrechtenSerializer
-from drc.datamodel.models.gebruiksrechten import Gebruiksrechten
-
-
-@conditional_retrieve()
-class GebruiksrechtenViewSet(
-    NotificationViewSetMixin,
-    CheckQueryParamsMixin,
-    ListFilterByAuthorizationsMixin,
-    AuditTrailViewsetMixin,
-    viewsets.ModelViewSet,
-):
-    """
-    Opvragen en bewerken van GEBRUIKSRECHTen bij een INFORMATIEOBJECT.
-
-    create:
-    Maak een GEBRUIKSRECHT aan.
-
-    Voeg GEBRUIKSRECHTen toe voor een INFORMATIEOBJECT.
-
-    **Opmerkingen**
-      - Het toevoegen van gebruiksrechten zorgt ervoor dat de
-        `indicatieGebruiksrecht` op het informatieobject op `true` gezet wordt.
-
-    list:
-    Alle GEBRUIKSRECHTen opvragen.
-
-    Deze lijst kan gefilterd wordt met query-string parameters.
-
-    retrieve:
-    Een specifieke GEBRUIKSRECHT opvragen.
-
-    Een specifieke GEBRUIKSRECHT opvragen.
-
-    update:
-    Werk een GEBRUIKSRECHT in zijn geheel bij.
-
-    Werk een GEBRUIKSRECHT in zijn geheel bij.
-
-    partial_update:
-    Werk een GEBRUIKSRECHT relatie deels bij.
-
-    Werk een GEBRUIKSRECHT relatie deels bij.
-
-    destroy:
-    Verwijder een GEBRUIKSRECHT.
-
-    **Opmerkingen**
-      - Indien het laatste GEBRUIKSRECHT van een INFORMATIEOBJECT verwijderd
-        wordt, dan wordt de `indicatieGebruiksrecht` van het INFORMATIEOBJECT op
-        `null` gezet.
-    """
-
-    queryset = Gebruiksrechten.objects.all()
-    serializer_class = GebruiksrechtenSerializer
-    filterset_class = GebruiksrechtenFilter
-    lookup_field = "uuid"
-    notifications_kanaal = KANAAL_DOCUMENTEN
-    notifications_main_resource_key = "informatieobject"
-    permission_classes = (InformationObjectRelatedAuthScopesRequired,)
-    required_scopes = {
-        "list": SCOPE_DOCUMENTEN_ALLES_LEZEN,
-        "retrieve": SCOPE_DOCUMENTEN_ALLES_LEZEN,
-        "create": SCOPE_DOCUMENTEN_AANMAKEN,
-        "destroy": SCOPE_DOCUMENTEN_ALLES_VERWIJDEREN,
-        "update": SCOPE_DOCUMENTEN_BIJWERKEN,
-        "partial_update": SCOPE_DOCUMENTEN_BIJWERKEN,
-    }
-    audit = AUDIT_DRC
-    audittrail_main_resource_key = "informatieobject"
