+++ src/drc/datamodel/admin.py
@@ -1,9 +1,6 @@
 from django.contrib import admin
-from django.utils.translation import gettext_lazy as _
 
 from privates.admin import PrivateMediaMixin
-
-from drc.datamodel.forms import VerzendingForm
 
 from .models import (
     BestandsDeel,
@@ -11,7 +8,6 @@ from .models import (
     EnkelvoudigInformatieObjectCanonical,
     Gebruiksrechten,
     ObjectInformatieObject,
-    Verzending,
 )
 
 
@@ -71,92 +67,3 @@ class BestandsDeelAdmin(PrivateMediaMixin, admin.ModelAdmin):
     list_display = ("__str__", "informatieobject", "volgnummer", "voltooid")
     list_filter = ("informatieobject",)
     private_media_fields = ("inhoud",)
-
-
-@admin.register(Verzending)
-class VerzendingAdmin(admin.ModelAdmin):
-    form = VerzendingForm
-
-    list_display = (
-        "uuid",
-        "aard_relatie",
-        "contactpersoonnaam",
-        "verzenddatum",
-        "ontvangstdatum",
-    )
-    list_filter = (
-        "aard_relatie",
-        "informatieobject",
-    )
-    ordering = (
-        "-verzenddatum",
-        "-ontvangstdatum",
-    )
-    search_fields = (
-        "contactpersoonnaam",
-        "uuid",
-    )
-    raw_id_fields = ("informatieobject",)
-
-    readonly_fields = ("uuid",)
-
-    fieldsets = (
-        (
-            _("Algemeen"),
-            {
-                "fields": (
-                    "uuid",
-                    "aard_relatie",
-                    "toelichting",
-                    "verzenddatum",
-                    "ontvangstdatum",
-                    "betrokkene",
-                    "informatieobject",
-                ),
-            },
-        ),
-        (
-            _("Contactpersoon"),
-            {
-                "fields": (
-                    "contact_persoon",
-                    "contactpersoonnaam",
-                ),
-            },
-        ),
-        (
-            _("Afwijkend binnenlands correspondentieadres verzending"),
-            {
-                "fields": (
-                    "binnenlands_correspondentieadres_huisletter",
-                    "binnenlands_correspondentieadres_huisnummer",
-                    "binnenlands_correspondentieadres_huisnummer_toevoeging",
-                    "binnenlands_correspondentieadres_naam_openbare_ruimte",
-                    "binnenlands_correspondentieadres_postcode",
-                    "binnenlands_correspondentieadres_woonplaats",
-                ),
-            },
-        ),
-        (
-            _("Afwijkend buitenlands correspondentieadres verzending"),
-            {
-                "fields": (
-                    "buitenlands_correspondentieadres_adres_buitenland_1",
-                    "buitenlands_correspondentieadres_adres_buitenland_2",
-                    "buitenlands_correspondentieadres_adres_buitenland_3",
-                    "buitenlands_correspondentieadres_land_postadres",
-                ),
-            },
-        ),
-        (
-            _("Afwijkend correspondentie postadres verzending"),
-            {
-                "fields": (
-                    "buitenlands_correspondentiepostadres_postbus_of_antwoord_nummer",
-                    "buitenlands_correspondentiepostadres_postadres_postcode",
-                    "buitenlands_correspondentiepostadres_postadrestype",
-                    "buitenlands_correspondentiepostadres_woonplaats",
-                ),
-            },
-        ),
-    )
