+++ src/drc/datamodel/constants.py
@@ -3,8 +3,6 @@ from django.utils.translation import ugettext_lazy as _
 from djchoices import ChoiceItem, DjangoChoices
 from vng_api_common import constants
 
-POSTAL_CODE_LENGTH = 6
-
 
 class Statussen(DjangoChoices):
     in_bewerking = ChoiceItem(
@@ -66,13 +64,3 @@ class ObjectInformatieObjectTypes(DjangoChoices):
     besluit = constants.BESLUIT_CHOICE
     zaak = constants.ZAAK_CHOICE
     verzoek = constants.VERZOEK_CHOICE
-
-
-class AfzenderTypes(DjangoChoices):
-    afzender = ChoiceItem("afzender", _("Afzender"))
-    geadresseerde = ChoiceItem("geadresseerde", _("Geadresseerde"))
-
-
-class PostAdresTypes(DjangoChoices):
-    antwoordnummer = ChoiceItem("antwoordnummer", _("Antwoordnummer"))
-    postbusnummer = ChoiceItem("postbusnummer", _("Postbusnummer"))
