--- src/drc/datamodel/models/__init__.py
@@ -1,9 +1,0 @@
-from .bestandsdeel import BestandsDeel  # noqa
-from .enkelvoudig_informatieobject import (  # noqa
-    EnkelvoudigInformatieObject,
-    EnkelvoudigInformatieObjectCanonical,
-)
-from .gebruiksrechten import Gebruiksrechten  # noqa
-from .informatieobject import InformatieObject  # noqa
-from .object_informatieobject import ObjectInformatieObject  # noqa
-from .verzending import Verzending  # noqa
