+++ src/drc/datamodel/validators.py
@@ -3,7 +3,7 @@ from datetime import date
 from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext_lazy as _
 
-from .constants import POSTAL_CODE_LENGTH, Statussen
+from drc.datamodel.constants import Statussen
 
 
 def validate_status(status: str = None, ontvangstdatum: date = None, instance=None):
@@ -39,28 +39,3 @@ def validate_status(status: str = None, ontvangstdatum: date = None, instance=No
                 )
             }
         )
-
-
-def validate_postal_code(value):
-    """
-    Validates a string is an correct dutch postal code.
-    """
-    if len(value) != POSTAL_CODE_LENGTH:
-        raise ValidationError(
-            _("Postcode moet %s tekens lang zijn.") % POSTAL_CODE_LENGTH,
-            code="invalid-length",
-        )
-
-    postal_digits = value[:4]
-
-    if not all(digit.isdigit() for digit in postal_digits):
-        raise ValidationError(_("De eerste vier karakters dienen een cijfer te zijn."))
-    elif not int(postal_digits) >= 1000 or not int(postal_digits) <= 9999:
-        raise ValidationError(
-            _("Alleen cijfers tussen de 1000 en 9999 zijn toegestaan")
-        )
-
-    postal_letters = value[4:]
-
-    if not all(letter.isalpha() for letter in postal_letters):
-        raise ValidationError(_("De laatste twee karakters dienen een letter te zijn."))
